article-generator/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
requests==2.31.0
